+++ crates/pipeline_manager/migrations/V15__program_compilation_profile_nullable.sql
@@ -1,0 +1,3 @@
+-- The program compilation profile can now also not be specified,
+-- in which case the compiler default compilation profile is used.
+ALTER TABLE program ALTER COLUMN compilation_profile DROP NOT NULL;