+++ crates/pipeline_manager/src/pipeline_automata.rs
@@ -744,7 +744,7 @@ mod test {
                 "program desc",
                 "ignored",
                 &ProgramConfig {
-                    profile: CompilationProfile::Unoptimized,
+                    profile: Some(CompilationProfile::Unoptimized),
                 },
                 None,
             )