+++ web-console/src/lib/services/manager/models/CompilationProfile.ts
@@ -3,10 +3,10 @@
 /* tslint:disable */
 /* eslint-disable */
 /**
- * Argument to `cargo build --profile <>` passed to the rust compiler
- *
- * Note that this is a hint to the backend, and can be overriden by
- * the Feldera instance depending on the administrator configuration.
+ * Enumeration of possible compilation profiles that can be passed to the Rust compiler
+ * as an argument via `cargo build --profile <>`. A compilation profile affects among
+ * other things the compilation speed (how long till the program is ready to be run)
+ * and runtime speed (the performance while running).
  */
 export enum CompilationProfile {
   DEV = 'dev',